Validate on write
