+++ cmd/list.go
@@ -31,7 +31,7 @@ func list(cmd *cobra.Command, args []string) error {
 		return errors.New("bucket not set")
 	}
 	service := strings.ToLower(args[0])
-	if err := validateService(service); err != nil {
+	if err := store.validateService(service); err != nil {
 		return errors.Wrap(err, "Failed to validate service")
 	}
