+++ cmd/write.go
@@ -31,12 +31,12 @@ func init() {
 
 func write(cmd *cobra.Command, args []string) error {
 	service := strings.ToLower(args[0])
-	if err := validateService(service); err != nil {
+	if err := store.validateService(service); err != nil {
 		return errors.Wrap(err, "Failed to validate service")
 	}
 
 	key := strings.ToLower(args[1])
-	if err := validateKey(key); err != nil {
+	if err := store.validateKey(key); err != nil {
 		return errors.Wrap(err, "Failed to validate key")
 	}
