+++ store/store.go
@@ -2,7 +2,11 @@ package store
 
 import (
 	"errors"
+	"fmt"
+	"regexp"
 	"time"
+
+	multierror "github.com/hashicorp/go-multierror"
 )
 
 var (
@@ -15,12 +19,35 @@ type Secrets struct {
 	Meta    *SecretsMetadata
 }
 
+var validKeyFormat = regexp.MustCompile(`^[A-Za-z0-9-_]+$`)
+var validServiceFormat = regexp.MustCompile(`^[A-Za-z0-9-_]+$`)
+
+func validateService(service string) error {
+	if !validServiceFormat.MatchString(service) {
+		return fmt.Errorf("Failed to validate service name '%s'.  Only alphanumeric, dashes, and underscores are allowed for service names", service)
+	}
+	return nil
+}
+
+func validateKey(key string) error {
+	if !validKeyFormat.MatchString(key) {
+		return fmt.Errorf("Failed to validate key name '%s'.  Only alphanumeric, dashes, and underscores are allowed for key names", key)
+	}
+	return nil
+}
+
 // A secret without any metadata
 type RawSecrets map[string]string
 
 func (r *RawSecrets) Validate() error {
-	// TODO; need to move validateKey here
-	panic("not implemented")
+	var result error
+	for k, _ := range *r {
+		err := validateKey(k)
+		if err != nil {
+			result = multierror.Append(result, err)
+		}
+	}
+	return result
 }
 
 type SecretsMetadata struct {
